pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["gym_mkds"]

[project]
name = "gym_mkds"
version = "0.2.1"
dependencies = [
  "gymnasium",
  "py-desmume-mkds>=0.0.15",
  "opencv-python>=4.13.0.92",
  "numpy",
  "trimesh==4.11.3",
  "scipy==1.17.1",
  "pre-commit",
]
